# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST para el sistema de monitoreo de instancias Proxmox El Centinela.\n**Flujo de autenticación**: Login → `/auth/login` → obtener `jwtTemporal` → `/auth/2fa/verify` → obtener `accessToken` → usar en el botón **Authorize** de esta UI.",
        "title": "El Centinela API",
        "contact": {
            "name": "Soporte El Centinela"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/account/password": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Valida la contraseña actual y aplica la nueva. Reglas de complejidad: entre 8 y 12 caracteres, al menos una mayúscula, un número y un carácter especial (!@#$%^\u0026*-_=+). Si la contraseña era temporal (`cambioContrasenaRequerido=true`), este cambio limpia ese flag y desbloquea el acceso al resto de la plataforma.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mi cuenta"
                ],
                "summary": "Cambiar mi contraseña",
                "parameters": [
                    {
                        "description": "Contraseña actual y nueva",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ports.CambiarContrasenaInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Formato inválido, contraseña actual incorrecta, nueva igual a la actual o no cumple las reglas de complejidad",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "PASSWORD_CHANGE_REQUIRED — solo este endpoint y logout son accesibles mientras el flag esté activo",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/account/profile": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Devuelve el perfil completo del usuario autenticado, incluyendo sus instancias Proxmox permitidas.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mi cuenta"
                ],
                "summary": "Obtener mi perfil",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ports.UsuarioDetalleDTO"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Permite al usuario modificar su propio `nombreCompleto` y/o `emailUsuario`. No permite cambiar rol ni organización.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mi cuenta"
                ],
                "summary": "Actualizar mi perfil",
                "parameters": [
                    {
                        "description": "Datos a actualizar",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ports.ActualizarPerfilInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ports.UsuarioResumenDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Email ya en uso",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/audit": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Devuelve el log de auditoría de la organización con soporte de filtros, paginación y ordenamiento. Solo accesible por ADMIN.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auditoría (Admin)"
                ],
                "summary": "Listar log de auditoría",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filtrar por UUID de usuario",
                        "name": "usuarioId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrar por tipo de acción (ej: LOGIN, CREAR_USUARIO)",
                        "name": "accion",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrar por ID de instancia Proxmox",
                        "name": "instanciaId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrar por resultado: EXITO o FALLA",
                        "name": "resultado",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Fecha desde (YYYY-MM-DD)",
                        "name": "desde",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Fecha hasta (YYYY-MM-DD)",
                        "name": "hasta",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Número de página (default 1)",
                        "name": "pagina",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Registros por página (default 50, máx 200)",
                        "name": "tamano",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Campo de ordenamiento: fechaHora (default), accion, resultado",
                        "name": "ordenarPor",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Dirección: desc (default) o asc",
                        "name": "direccion",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ports.PaginaAuditoria"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/admin/audit/export": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Exporta el log de auditoría de la organización. Aplica los mismos filtros que el listado. Retorna el archivo como adjunto descargable. Solo accesible por ADMIN.",
                "produces": [
                    "text/csv",
                    "application/json"
                ],
                "tags": [
                    "Auditoría (Admin)"
                ],
                "summary": "Exportar log de auditoría",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filtrar por UUID de usuario",
                        "name": "usuarioId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrar por tipo de acción",
                        "name": "accion",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrar por ID de instancia",
                        "name": "instanciaId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "EXITO o FALLA",
                        "name": "resultado",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Fecha desde (YYYY-MM-DD)",
                        "name": "desde",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Fecha hasta (YYYY-MM-DD)",
                        "name": "hasta",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Formato de exportación: csv (default) o json",
                        "name": "formato",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Archivo CSV o JSON descargable",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/admin/users": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Devuelve todos los usuarios de la organización del admin autenticado, con un resumen (`summary`) de totales por rol. Soporta filtros opcionales: `?rol=ADMIN|OPERATOR`, `?activo=true|false`, `?buscar=texto` (nombre o email).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Usuarios (Admin)"
                ],
                "summary": "Listar usuarios de la organización",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filtrar por rol: ADMIN o OPERATOR",
                        "name": "rol",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filtrar por estado: true o false",
                        "name": "activo",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Buscar por nombre o email (case-insensitive)",
                        "name": "buscar",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ports.ListaUsuariosResult"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Crea un nuevo usuario con contraseña temporal generada automáticamente. La contraseña se envía al usuario por email. El usuario deberá cambiarla en su primer login.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Usuarios (Admin)"
                ],
                "summary": "Crear usuario",
                "parameters": [
                    {
                        "description": "Datos del nuevo usuario",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ports.CrearUsuarioInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/ports.CrearUsuarioResult"
                        }
                    },
                    "400": {
                        "description": "Datos inválidos",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Email o username ya registrado",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/auth/2fa/verify": {
            "post": {
                "security": [
                    {
                        "BearerPreAuth": []
                    }
                ],
                "description": "Recibe el código de 6 dígitos del autenticador. Si es correcto emite el `accessToken` (8h) y el `refreshToken` (30 días) con 2FA completado. Funciona tanto para la primera vinculación como para logins posteriores.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Autenticación 2FA"
                ],
                "summary": "Verificar código TOTP → obtener tokens definitivos",
                "parameters": [
                    {
                        "description": "Código TOTP de 6 dígitos",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.verificarTotpRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ports.TokenResult"
                        }
                    },
                    "400": {
                        "description": "Código inválido (no tiene 6 dígitos)",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Código TOTP incorrecto o ya utilizado (anti-replay)",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "http.ErrorResponse": {
            "type": "object",
            "properties": {
                "errorCode": {
                    "type": "string",
                    "example": "INVALID_REQUEST"
                },
                "message": {
                    "type": "string",
                    "example": "Mensaje descriptivo del error."
                }
            }
        },
...
        "ports.CambiarContrasenaInput": {
            "type": "object",
            "required": [
                "contrasenaActual",
                "contrasenaNueva"
            ],
            "properties": {
                "contrasenaActual": {
                    "type": "string"
                },
                "contrasenaNueva": {
                    "type": "string",
                    "maxLength": 12,
                    "minLength": 8
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Token de acceso definitivo. Formato: **Bearer \u0026lt;accessToken\u0026gt;**. Obtené el token completando el flujo: `POST /auth/login` → `POST /auth/2fa/verify`.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BearerPreAuth": {
            "description": "JWT temporal pre-2FA. Formato: **Bearer \u0026lt;jwtTemporal\u0026gt;**. Obtené el token de `POST /auth/login`.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}